// src/components/interface.ts
export enum MainUrls {
  index = "/",
  MainCategories = "categories",
  MainCategoriesId = "MainCategoriesID",
  MainTrending = "trending",
  MainDisclosure = "disclosure",
  Favorites = "Favorites",
  FilmId = "FilmId",
}

export enum AuthGuardUrls {
  index = "/auth",
  logIn = "login",
  register = "register",
}